refactor(desktop): extract daily review ipc
